refactor(api): extract shared ResponseModel into fields/base.py
Consolidate the identical ResponseModel definition that was duplicated
across 7 files into a single canonical location at fields/base.py.
