+++ api/fields/base.py
@@ -1,0 +1,13 @@
+from __future__ import annotations
+
+from pydantic import BaseModel, ConfigDict
+
+
+class ResponseModel(BaseModel):
+    model_config = ConfigDict(
+        from_attributes=True,
+        extra="ignore",
+        populate_by_name=True,
+        serialize_by_alias=True,
+        protected_namespaces=(),
+    )
