+++ api/fields/member_fields.py
@@ -4,7 +4,9 @@ from datetime import datetime
 
 from flask_restx import fields
 from graphon.file import helpers as file_helpers
-from pydantic import BaseModel, ConfigDict, computed_field, field_validator
+from pydantic import computed_field, field_validator
+
+from fields.base import ResponseModel
 
 simple_account_fields = {
     "id": fields.String,
@@ -25,16 +27,6 @@ def _build_avatar_url(avatar: str | None) -> str | None:
     if avatar.startswith(("http://", "https://")):
         return avatar
     return file_helpers.get_signed_file_url(avatar)
-
-
-class ResponseModel(BaseModel):
-    model_config = ConfigDict(
-        from_attributes=True,
-        extra="ignore",
-        populate_by_name=True,
-        serialize_by_alias=True,
-        protected_namespaces=(),
-    )
 
 
 class SimpleAccount(ResponseModel):
